+++ src/tests/TestCase.php
@@ -10,11 +10,16 @@ abstract class TestCase extends BaseTestCase
 
     protected static $client;
     protected static $expireCassetteDays = 180;
+    protected static $prodApiKey;
+    protected static $testApiKey;
 
     protected function setUp(): void
     {
         parent::setUp();
 
         $this->withoutVite();
+
+        self::$prodApiKey = getenv('EASYPOST_PROD_API_KEY');
+        self::$testApiKey = getenv('EASYPOST_TEST_API_KEY');
     }
 }